FIX:: useOutsideClick hook
